Disposing old binding
In order to prevent a potential memory leak, the old
AutoCompletionBinding is disposed. As this is done in a lambda
expression, a reference must be stored to dispose the old one. This is
done using an AtomicReference. If this is a suitable approach is open
for discussion.
In order to prevent simultaneous changes of the binding reference, the
whole action is synchronized.