Updating docker support
The build process of the docker image has been optimized to cache the
maven dependencies and therefore eliminate the need of always
redownloading the dependencies.
Furthermore, documentation has been added that explains, how to use the
docker image and how to build it yourself.
Finally, the properties files have been checked and differentiated into
different files in order to allow clean separation of profiles. For now
the server has the following profiles:
- default
- dev
- prod

the client has no prod profile so far. This might change in the future.